src/clip-search-engine/requirements.txt
pinecone-client
fiftyone
torch
torchvision
Pillow